+++ tests/tests_django_peertube_runner_connector/views/video/test_transcode_video.py
@@ -102,5 +102,5 @@ class TranscodeVideoAPITest(TestCase):
             self.assertEqual(
                 runner_job.payload["input"]["videoFileUrl"],
                 "http://testserver/api/v1/runners/jobs/"
-                f"files/videos/{created_video.uuid}/max-quality",
+                f"files/videos/{created_video.uuid}/{runner_job.uuid}/max-quality",
             )